+++ components/templates/PostDetail.tsx
@@ -1,5 +1,5 @@
 import { Layout } from '@components/layouts';
-import { usePostQuery } from '@graphql/post.graphql';
+import { usePostQuery } from '@graphql/queries/post.graphql';
 import Image from 'next/image';
 import Link from 'next/link';
 import React from 'react';
