+++ lib/graphql/mutations/auth.graphql
