+++ lib/graphql/mutations/collection.graphql
@@ -1,23 +1,3 @@
-query Collection($id: Int!) {
-  collection(id: $id) {
-    id
-    postCount
-    preview
-    name
-    description
-    createdAt
-    author {
-      id
-      username
-      firstName
-      lastName
-    }
-    posts {
-      id
-    }
-  }
-}
-
 mutation editCollection(
   $name: String!
   $description: String!
