+++ lib/graphql/mutations/user.graphql
@@ -1,0 +1,7 @@
+mutation deleteUser($username: String!, $password: String!) {
+  deleteUser(input: { username: $username, password: $password })
+}
+
+mutation updateProfile($input: UpdateProfileInput!) {
+  updateProfile(input: $input)
+}
