+++ lib/graphql/queries/collection.graphql
@@ -1,0 +1,19 @@
+query Collection($id: Int!) {
+  collection(id: $id) {
+    id
+    postCount
+    preview
+    name
+    description
+    createdAt
+    author {
+      id
+      username
+      firstName
+      lastName
+    }
+    posts {
+      id
+    }
+  }
+}
