+++ lib/graphql/queries/place.graphql
