+++ lib/graphql/queries/user.graphql
@@ -49,14 +49,6 @@ query user($input: UserInput!) {
   }
 }
 
-mutation deleteUser($username: String!, $password: String!) {
-  deleteUser(input: { username: $username, password: $password })
-}
-
-mutation updateProfile($input: UpdateProfileInput!) {
-  updateProfile(input: $input)
-}
-
 query Me {
   me {
     id
