+++ pages/create/post.tsx
@@ -2,8 +2,8 @@ import { Input, Separator } from '@components/elements';
 import Button from '@components/elements/buttons/Button';
 import { Layout } from '@components/layouts';
 import Search from '@components/modules/Search';
-import { useCreatePostMutation } from '@graphql/post.graphql';
-import { useMeQuery } from '@graphql/user.graphql';
+import { useCreatePostMutation } from '@graphql/mutations/post.graphql';
+import { useMeQuery } from '@graphql/queries/user.graphql';
 import { RedirectAnonymous } from '@lib/functions/ServerSideProps';
 import { motion } from 'framer-motion';
 import { useRouter } from 'next/router';
