+++ pages/settings/notifications.tsx
@@ -1,6 +1,6 @@
 import { CheckboxElement } from '@components/elements';
 import SettingsLayout from '@components/layouts/Settings';
-import { useUpdateProfileMutation } from '@graphql/user.graphql';
+import { useUpdateProfileMutation } from '@graphql/mutations/user.graphql';
 import { RedirectAnonymous } from '@lib/functions/ServerSideProps';
 import React from 'react';
 import { useTranslation } from 'react-i18next';
