+++ pages/user/[username]/map.tsx
@@ -2,8 +2,8 @@ import { ApolloClient, createHttpLink, InMemoryCache } from '@apollo/client';
 import UserLayout from '@components/layouts/User';
 import MapStyleMenu from '@components/modules/map/MapStyleMenu';
 import Card from '@components/modules/userPage/Card';
-import { PostsDocument } from '@graphql/post.graphql';
-import { UserDocument, UserQuery } from '@graphql/user.graphql';
+import { PostsDocument } from '@graphql/queries/post.graphql';
+import { UserDocument, UserQuery } from '@graphql/queries/user.graphql';
 import {
   GetCookieFromServerSideProps,
   IsJwtValid,
